% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_screeners.R
\name{apply_screeners}
\alias{apply_screeners}
\title{Apply growth screeners to child data}
\usage{
apply_screeners(
  txt = "",
  session = "",
  loc = "",
  format = "1.0",
  ynames = c("hgt", "wgt", "hdc"),
  na.omit = TRUE,
  ...
)
}
\arguments{
\item{txt}{A JSON string, URL or file}

\item{session}{Alternative to \code{txt}. Session key where input data is uploaded.}

\item{loc}{Alternative to \code{txt}. Location where input data is uploaded.}

\item{format}{String. JSON data schema version number. There are currently
three schemas supported: \code{"1.0"}, \code{"1.1"} and \code{"2.0"}. Formats \code{"1.0"} and
\code{"1.1"} are included for backward compatibility only. Use \code{format = "2.0"}
for new applications.}

\item{ynames}{Character vector identifying the measures to be
screened. By default, \code{ynames = c("hgt", "wgt", "hdc")}.}

\item{na.omit}{A logical indicating whether records with a missing
\code{x} (age) or \code{y} (yname) should be removed. Defaults to
\code{TRUE}.}

\item{...}{Ignored}
}
\description{
Apply growth screeners to child data
}
\note{
\code{apply_screeners()} supersedes \code{screen_growth()} and \code{screen_curves()}.
}
\examples{
fn <- system.file("testdata", "client3.json", package = "james")
apply_screeners(fn)
\dontrun{
# first upload, then screen
host <- "http://localhost"
r1 <- jamesclient::upload_txt(fn, host = host)
location <- jamesclient::get_url(r1, "location")
location
apply_screeners(loc = location)
}
}
