% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ddomain.R
\name{calculate_ddomain}
\alias{calculate_ddomain}
\title{Calculates the domain specific D-score and DAZ for each visit}
\usage{
calculate_ddomain(
  txt = "",
  session = "",
  format = "3.1",
  append = c("ddi", "gs1"),
  set = "GFCLS",
  domain = NULL,
  loc = "",
  ...
)
}
\arguments{
\item{txt}{A JSON string, URL, or file with BDS data in JSON format. Data
should conform to the BDS JGZ 3.2.5 specification.}

\item{session}{Optional session key if data is already uploaded to \code{sitehost}.}

\item{format}{JSON schema version, e.g., \code{"3.0"}. Used when uploading.}

\item{append}{Optional vector of strings indicating which instrument to base
D-score calculations on. Currently supports \code{ddi} and \code{gs1}. Requires JSON
schema V3.0 or later.}

\item{set}{String with the name of the domainset}

\item{domain}{Vector of strings with names of specific domains, by default
all domains in the set are returned.}

\item{loc}{Deprecated. Use \code{session} instead.}

\item{...}{Ignored}
}
\value{
A list of data.frames.
}
\description{
The function \code{draw_chart()} plots individual data on the growth chart.
}
\examples{
fn <- system.file("testdata", "Laura_S.json", package = "james")
df <- calculate_dscore(txt = fn)
head(df, 4)
}
\author{
Iris Eekhout 2025
}
\keyword{server}
