% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dscore.R
\name{calculate_dscore}
\alias{calculate_dscore}
\title{Calculates the D-score and DAZ for each visit}
\usage{
calculate_dscore(
  txt = "",
  session = "",
  format = "3.1",
  append = c("ddi", "gs1"),
  key = NULL,
  population = NULL,
  output = c("table", "last_visit", "last_dscore"),
  loc = "",
  ...
)
}
\arguments{
\item{txt}{A JSON string, URL or file}

\item{session}{Optional session key if data is already uploaded to \code{sitehost}.}

\item{format}{String. JSON data schema version number. There are currently
three schemas supported: \code{"1.0"}, \code{"1.1"}, \code{"2.0"}, \code{"3.0"} and \code{"3.1"}.
Formats \code{"1.0"} and \code{"1.1"} are included for backward compatibility only.
Use \code{format = "3.1"} for new applications.}

\item{append}{Optional vector of strings indicating which instrument to base
D-score calculations on. Currently supports \code{ddi} and \code{gs1}. Requires JSON
schema V3.0 or later.}

\item{key}{String. They key identifies 1) the difficulty estimates
pertaining to a particular Rasch model, and 2) the prior mean and standard
deviation of the prior distribution for calculating the D-score.
The default key \code{NULL} sets \code{key = "gsed2510"}.
View \code{builtin_keys} for an overview of the available keys.}

\item{population}{String. The name of the reference population to calculate
DAZ.
Use \code{with(builtin_references, table(key, population))} to see which
built-in references are available for \code{key - population} combinations.
If not specified, the function set the default population as
\code{builtin_keys$base_population[key == builtin_keys$key]}.}

\item{output}{A string, either \code{"table"}, \code{"last_visit"} or '\code{"last_dscore"}
specifying the result. The default \code{"table"} returns with columns: \code{"date"}
(date), \code{"x"} (age), \code{"y"} (D-score) and \code{"z"} (DAZ). The number of rows
equals to the number of visits. If \code{output} equals \code{"last_visit"} the
function returns only the last row. If \code{output} equals \code{"last_dscore"} the
function returns only the D-score from the last row.}

\item{loc}{Deprecated. Use \code{session} instead.}

\item{...}{Ignored}
}
\value{
A table, row or scalar.
}
\description{
The function \code{calculate_dscore()} calculate the D-score for early child
development.
}
\examples{
fn <- system.file("testdata", "Laura_S.json", package = "james")
df <- calculate_dscore(txt = fn)
head(df, 4)
}
\author{
Stef van Buuren 2020-2026
}
\keyword{server}
