% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.R
\name{chart}
\alias{chart}
\title{Draw growth chart with individual data}
\usage{
chart(location = NULL, chartcode = NULL, curve_interpolation = TRUE,
  ...)
}
\arguments{
\item{location}{A url that points to the server location where the
parsed data are stored.}

\item{chartcode}{The code of the requested growth chart. If not
specified, the server will automatically plot child height for
the most recent age period.}

\item{curve_interpolation}{A logical indicating whether curve interpolation shoud be applied.}

\item{\dots}{Not used}
}
\value{
tbd
}
\description{
This function expect an input key to an object of class
\linkS4class{individual}, and calculates the individual data on the
requested growth chart.
}
\seealso{
\linkS4class{individual},
\code{\link[groeidiagrammen]{select_chart}}
\code{\link[groeidiagrammen]{draw_plot}}
}
\author{
Stef van Buuren 2019
}
\keyword{server}
