% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ind_chartcodelist.R
\name{convert_ind_chartcodelist}
\alias{convert_ind_chartcodelist}
\title{Derive chartcode list from data}
\usage{
convert_ind_chartcodelist(location = NULL, chartcode = NULL)
}
\arguments{
\item{location}{A url that points to the server location where the
data from a previous request to \code{convert_bds_ind()} are stored.}

\item{chartcode}{The code of the requested growth chart. If not
specified, the server will automatically plot child height for
the most recent age period.}
}
\value{
A list with chart codes, produced by \code{chartbox::parse_chartcode}
}
\description{
The function load individual data into, calculates the
chartcode and returns a list of parsed chartcode. The
function is called at initialization to automate the
selection of the chart according to the individual data.
}
\seealso{
\code{\link[chartbox]{parse_chartcode}}
}
\author{
Stef van Buuren 2019
}
\keyword{server}
