% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_list.R
\name{custom_list}
\alias{custom_list}
\title{Provides a Screen growth curves according to JGZ guidelines}
\usage{
custom_list(txt = NULL, loc = NULL)
}
\arguments{
\item{txt}{A JSON string, URL or file, with the data (in JSON
format) to be uploaded. The variable specification are expected
to be according specification
\href{https://www.ncj.nl/themadossiers/informatisering/basisdataset/documentatie/?cat=12}{BDS
 JGZ 3.2.5}, and converted to JSON.}

\item{loc}{Alternative to \code{txt}. Location where data is uploaded
and processed, for example, as obtained by \code{upload_txt()}.}
}
\value{
A table with screening results

A list with custom parts
}
\description{
Provides a Screen growth curves according to JGZ guidelines
}
\examples{
\dontrun{
fn <- system.file("extdata", "smocc", "Laura_S.json", package = "jamestest")

# first upload, then create custom list
r1 <- upload_txt(fn)
loc <- jamesclient::get_url(r1, "location")
list1 <- custom_list(loc = loc)

# upload & screen
list2 <- custom_list(fn)
identical(list1, list2)
}
}
