% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcat_start.R
\name{dcat_start}
\alias{dcat_start}
\title{Calculates the start item from age for an adaptive test to measure D-score.}
\usage{
dcat_start(
  txt = "",
  instrument = "gs1",
  key = NULL,
  population = NULL,
  p = 50,
  session = "",
  format = "3.1",
  ...
)
}
\arguments{
\item{txt}{A JSON string, URL or file}

\item{instrument}{character vector with instrument names to use to select
items}

\item{key}{String. They key identifies 1) the difficulty estimates
pertaining to a particular Rasch model, and 2) the prior mean and standard
deviation of the prior distribution for calculating the D-score.
The default key \code{NULL} sets \code{key = "gsed2510"}.
View \code{dscore::builtin_keys} for an overview of the available keys.}

\item{population}{String. Name of the reference population}

\item{p}{percentage to pass the item, difficulty in percentile units}

\item{session}{Optional session key if data is already uploaded to \code{sitehost}.}

\item{format}{String. JSON data schema version number. There are currently
three schemas supported: \code{"1.0"}, \code{"1.1"}, \code{"2.0"}, \code{"3.0"} and \code{"3.1"}.
Formats \code{"1.0"} and \code{"1.1"} are included for backward compatibility only.
Use \code{format = "3.1"} for new applications.}

\item{...}{Ignored}
}
\value{
A string of milestones.
}
\description{
Calculates the start item from age for an adaptive test to measure D-score.
}
\examples{
fn <- system.file("examples", "example_v3.1.json", package = "bdsreader")
dcat_start(txt = fn, p = 50)
}
\author{
Iris Eekhout 2025
}
