% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{draw_chart_bds-deprecated}
\alias{draw_chart_bds-deprecated}
\alias{draw_chart_bds}
\title{Convert bds-format data to individual and return growth chart}
\usage{
draw_chart_bds(
  txt = "",
  format = "1.0",
  chartcode = "",
  curve_interpolation = TRUE,
  selector = "chartcode",
  ...
)
}
\arguments{
\item{txt}{A JSON string, URL or file}

\item{format}{String. JSON data schema version number. There are currently
three schemas supported: \code{"1.0"}, \code{"1.1"}, \code{"2.0"} and \code{"3.0"}.
Formats \code{"1.0"} and \code{"1.1"} are included for backward compatibility only.
Use \code{format = "3.0"} for new applications.}

\item{chartcode}{A string with chart code}

\item{curve_interpolation}{A logical indicating whether curve
interpolation shoud be applied.}

\item{selector}{Legacy addition to solve a problem in jgzApp. See
\code{draw_chart} for interpretation. The default is set to
\code{"chartcode"}.}

\item{\dots}{For \code{draw_chart_bds}, additional parameter passed
down to \code{fromJSON(txt, ...)}, \code{new("xyz",... )} and
\code{new("bse",... )}. Useful parameters are \code{models = "bsmodel"} for setting the broken stick model, or \code{call = as.call(...)} for setting proper reference standards.}
}
\description{
The function \code{draw_chart_bds()} convert bds data into an object
of class individual, and then draws the
individual data on the requested growth chart.
Superseded by \code{\link[=draw_chart]{draw_chart()}}.
}
\examples{
fn <- system.file("testdata", "client3.json", package = "james")
g <- draw_chart_bds(txt = fn)
}
\keyword{server}
