% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_chart_bds.R
\name{draw_chart_bds}
\alias{draw_chart_bds}
\title{Convert bds-format data to individual and return growth chart}
\usage{
draw_chart_bds(txt = "", chartcode = "", curve_interpolation = TRUE, ...)
}
\arguments{
\item{txt}{A JSON string, URL or file}

\item{chartcode}{A string with chart code}

\item{curve_interpolation}{A logical indicating whether curve
interpolation shoud be applied.}

\item{\dots}{For \code{draw_chart_bds}, additional parameter passed
down to \code{fromJSON(txt, ...)}, \code{new("xyz",... )} and
\code{new("bse",... )}. Useful parameters are \code{models =
"bsmodel"} for setting the broken stick model, or \code{call =
as.call(...)} for setting proper reference standards.}
}
\description{
The function \code{draw_chart_bds()} convert bds data into an object
of class \linkS4class{individual}, and then draws the
individual data on the requested growth chart.
}
\examples{
fn <- system.file("testdata", "client3.json", package = "james")
g <- draw_chart_bds(txt = fn)
}
\keyword{server}
