% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_plot.R
\name{draw_plot}
\alias{draw_plot}
\title{Constructs and plots the A4 growth chart}
\usage{
draw_plot(
  individual,
  chartcode,
  curve_interpolation = TRUE,
  quiet = TRUE,
  con = NULL,
  dnr = NULL,
  period = numeric(0),
  nmatch = 0L,
  user_model = 2L,
  exact_sex = TRUE,
  exact_ga = FALSE,
  break_ties = TRUE,
  show_realized = FALSE,
  show_future = FALSE,
  clip = TRUE
)
}
\arguments{
\item{individual}{An S4 object of class \code{individual}
containing data of the individual}

\item{chartcode}{A string with chart code}

\item{curve_interpolation}{A logical indicating whether curve
interpolation shoud be applied.}

\item{quiet}{Logical indicating whether chart code should be
written to standard output. Default is \code{quiet = TRUE}.}

\item{con}{A connection on which the donor data reside. The
default (\code{NULL}) reads from \code{donorloader} package.}

\item{dnr}{A string with the name of the donor data
(currently available are \code{smocc}, \code{terneuzen},
\code{lollypop.preterm} or \code{lollypop.term})}

\item{period}{A vector of length 2 with left and right ages
(decimal age). If \code{length(period) == 0L}, then no curve
matching is done}

\item{nmatch}{Integer. Number of matches needed. When
\code{nmatch == 0L} no matches are sought.}

\item{user_model}{Model number (1-4), indicating type of model the
user wants. See details.}

\item{exact_sex}{A logical indicating whether sex should be
matched exactly}

\item{exact_ga}{A logical indicating whether gestational age
should be matched exactly}

\item{break_ties}{A logical indicating whether ties should broken
randomly. The default (\code{TRUE}) breaks ties randomly.}

\item{show_realized}{A logical indicating whether the realized
growth of the target child should be drawn}

\item{show_future}{A logical indicating whether the predicted
growth of the target child should be drawn}

\item{clip}{A logical indicating whether clipping is needed}
}
\value{
A grid object
}
\description{
This function is a wrapper of \code{chartplotter::process_chart}
}
