% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_loc.R
\name{fetch_loc-deprecated}
\alias{fetch_loc-deprecated}
\alias{fetch_loc}
\title{Uploads, parses, converts and stores data on the server for further processing}
\usage{
fetch_loc(txt = "", ...)
}
\arguments{
\item{txt}{A JSON string, URL or file}

\item{...}{Ignored}
}
\value{
An object of class \code{target}. Basically a list with elements \code{psn}
(persondata) and \code{xyz} (timedata).
}
\description{
Uploads JSON data that adhere to the BDS-format, parses its
contents, converts it to an internal JAMES object of class \code{target},
and stores the result on the server for further processing.
The function is useful for caching input data over multiple requests to
\code{OpenCPU}. The cached data feed into other JAMES functions by means
of the \code{"loc"} argument. The server wipes the cached data after 2 hours.
}
\examples{
fn <- system.file("testdata", "client3.json", package = "james")
p <- fetch_loc(fn)
}
\seealso{
\code{\link[bdsreader:read_bds]{bdsreader::read_bds()}}
\code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}}
}
\author{
Stef van Buuren 2021
}
\keyword{server}
