% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_session_local.R
\name{get_session_object}
\alias{get_session_object}
\title{pass in the url to the RDS representation of a openCPU session object, get the object}
\usage{
get_session_object(session, object = ".val")
}
\arguments{
\item{session}{Character, the session}

\item{object}{Character, e.g. ".val"}
}
\description{
useful to programmatically access openCPU session object stored in character variables etc.
}
\examples{
\dontrun{
get_session_object(session = "x02a93ec661121", object = ".val")
}
}
