% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j_get.R
\name{j_get}
\alias{j_get}
\title{Get data from archive}
\usage{
j_get(type, version, scenario = .j_root$active_scenario,
  project = .j_root$active_project, what = c("data", "fig", "object"),
  index)
}
\description{
Get data from archive
}
\examples{
j_put(x = 10, type = "number", scenario = "Primary school", project = "Education")
j_get(type = "number")
j_put(x = 11)
j_get("number")
j_get("number", version = "2")
}
