% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j_init.R
\name{j_init}
\alias{j_init}
\title{Initialize James}
\usage{
j_init(file_name = "data.james", active_scenario, active_project)
}
\arguments{
\item{file_name}{File holding your archive. Defaults to "data.james".}

\item{active_scenario}{Scenario you want to work on.}

\item{active_project}{Project you want to work on.}
}
\description{
Initialize James
}
\examples{
j_init()
}
