% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j_put.R
\name{j_put}
\alias{j_put}
\title{Put data in archive.}
\usage{
j_put(x, type = "", doc = NA, scenario = .j_root$active_scenario,
  project = .j_root$active_project, ignore_if_duplicate = TRUE,
  activate_project_scenario = TRUE)
}
\description{
Put data in archive.
}
\examples{
j_put(x = 10, type = "number", scenario = "Primary school", project = "Education")
j_get(type = "number")
j_put(x = 11)
j_get("number")
j_get("number", version = "2")
}
