% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/james-objects.R
\docType{data}
\name{JFig}
\alias{JFig}
\title{R6 objects holding your data}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
JFig
}
\description{
R6 objects holding your data
}
\keyword{datasets}
