% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_charts.R
\name{list_charts}
\alias{list_charts}
\title{List available growth charts}
\usage{
list_charts(chartgrp = NULL, ...)
}
\arguments{
\item{chartgrp}{Optional. String chart group name, e.g. \code{chartgrp = "who"}.
If omitted, \code{list_charts()} return all charts groups.}

\item{\dots}{Not used}
}
\value{
A \code{data.frame} with eight columns:
\describe{
\item{\code{chartgrp}}{Chart group code}
\item{\code{chartcode}}{Unique chart code}
\item{\code{population}}{Population code: DS (Down Syndrome), HS (Hindostanic),
MA (Morrocan), NL (Dutch), PT (Preterm), TU (Turkish), WHOblue (WHO male),
WHOpink (WHO female)}
\item{\code{sex}}{Either \code{"male"}  or \code{"female"}}
\item{\code{design}}{Chart design A: 0-15m, B: 0-4y (WFH), C: 1-21y, D: 0-21y, E: 0-4y (WFA)}
\item{\code{side}}{Outcome measure: hgt (height), wgt (weight), hdc (head circumference),
wfh (weight for height), bmi (body mass index), front (multiple), back (multiple),
-hdc (back, no head circumference)}
\item{\code{language}}{Chart language: dutch}
\item{\code{week}}{Weeks of gestation 25-36, or missing (>= 37 weeks)}
}
}
\description{
List available growth charts
}
\examples{
# list all available charts
list_charts()
}
