% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_screeners.R
\name{list_screeners}
\alias{list_screeners}
\title{List the available growth screeners}
\usage{
list_screeners(ynames = c("hgt", "wgt", "hdc"), ...)
}
\arguments{
\item{ynames}{Character vector identifying the measures to be
screened. By default, \code{ynames = c("hgt", "wgt", "hdc")}.}

\item{\dots}{Used for authentication, passed down to \code{\link[growthscreener:list_screeners]{growthscreener::list_screeners()}}.}
}
\description{
List the available growth screeners
}
\examples{
head(list_screeners(ynames = "hgt"), 2)
}
