% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sds_chart.R
\name{make_sds_chart}
\alias{make_sds_chart}
\title{Make SDS charts}
\usage{
make_sds_chart(title = "This is a test", lat = NULL, lng = NULL)
}
\arguments{
\item{title}{label for the marker}

\item{lat}{lattitude coordinate}

\item{lng}{longintude coordinate}
}
\description{
Make SDS charts
}
