% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentiles_vwc.R
\name{percentiles_vwc}
\alias{percentiles_vwc}
\title{Provides additional information on percentiles of D-score and age in months
for the percentiles 2, 10, 50, 90 and 98. User either provides known van
Wiechen milestones, or information to calculate the recommended van Wiechen
milestones based on age and past milestones.}
\usage{
percentiles_vwc(
  txt = "",
  p = 90,
  n = 6,
  vwc = NULL,
  session = "",
  format = "1.0",
  loc = "",
  percentiles = FALSE,
  ...
)
}
\arguments{
\item{txt}{A JSON string, URL, or file with BDS data in JSON format. Data
should conform to the BDS JGZ 3.2.5 specification.}

\item{p}{Reference percentile indicating the expected probability of a
positive van Wiechen outcome given child age. Higher values correspond with
easier items.}

\item{n}{Number of van Wiechen items to suggest. By default returns all items
within set probability limits.}

\item{vwc}{Vector of string values matching van Wiechen milestones. Bypasses
calculations and directly supply the milestones of interest}

\item{session}{Optional session key if data is already uploaded to \code{sitehost}.}

\item{format}{JSON schema version, e.g., \code{"3.0"}. Used when uploading.}

\item{loc}{Deprecated. Use \code{session} instead.}

\item{...}{Ignored}
}
\value{
A table.
}
\description{
Provides additional information on percentiles of D-score and age in months
for the percentiles 2, 10, 50, 90 and 98. User either provides known van
Wiechen milestones, or information to calculate the recommended van Wiechen
milestones based on age and past milestones.
}
\examples{
fn <- system.file("testdata", "Laura_S.json", package = "james")
percentiles_vwc(txt = fn, p = 50, n = 3)
}
\author{
Arjan Huizing 2025
}
