% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomplot.R
\name{randomplot}
\alias{randomplot}
\title{Make a random plot}
\usage{
randomplot(n, dist = c("normal", "uniform"))
}
\arguments{
\item{n}{numer of random values}

\item{dist}{one of "normal" or "uniform".}
}
\description{
This function creates a random histogram plot.
}
