% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_site_ind.R
\name{render_site_ind}
\alias{render_site_ind}
\title{Render the site starting from BDS data}
\usage{
render_site_ind(location = NULL, chartcode = NULL,
  curve_interpolation = TRUE)
}
\arguments{
\item{location}{A url that points to the server location where the
data from a previous request to \code{convert_bds_ind()} are stored.}

\item{chartcode}{The code of the requested growth chart. If not
specified, the server will automatically plot child height for
the most recent age period.}

\item{curve_interpolation}{A logical indicating whether curve interpolation shoud be applied.}
}
\value{
A url pointing to the site
}
\description{
The function constructs the url that points to the site
with the individual data.
}
\seealso{
\linkS4class{individual},
\code{\link{draw_chart_ind}}
}
\author{
Stef van Buuren 2019
}
\keyword{server}
