% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_blend.R
\name{request_blend}
\alias{request_blend}
\title{Provides multiple outputs in one request}
\usage{
request_blend(
  txt = "",
  sitehost = "",
  session = "",
  blend = "standard",
  loc = "",
  ...
)
}
\arguments{
\item{txt}{A JSON string, URL, or file with BDS data in JSON format. Data
should conform to the BDS JGZ 3.2.5 specification.}

\item{sitehost}{The server that renders the site. Defaults to
\code{"http://localhost:8080"} if not specified.}

\item{session}{Optional session key if data is already uploaded to \code{sitehost}.}

\item{blend}{A string indicating the requested blend. The default (\code{"standard"})
returns the results of the standard end points that produces tables. Graphs
are currently not supported.}

\item{loc}{Deprecated. Use \code{session} instead.}

\item{...}{Ignored}
}
\value{
The default \code{blend = "standard"} return a list with the following
components:
\describe{
\item{\code{txt}}{String, file or URL with child data}
\item{\code{session}}{Session with uploaded child data}
\item{\code{child}}{Processed child level data}
\item{\code{time}}{Processed time level data}
\item{\code{screeners}}{Results from application of screeners to child data}
\item{\code{site}}{URL with personalised child data}
}
}
\description{
Function \code{request_blend()} acts as a one-stop-shop to obtain multiple
outputs through one request.
}
\examples{
\dontrun{
fn <- system.file("extdata", "bds_v2.0", "smocc", "Laura_S.json", package = "jamesdemodata")
results <- request_blend(txt = fn)
}
}
