% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{screen_curves-deprecated}
\alias{screen_curves-deprecated}
\alias{screen_curves}
\title{Screen growth curves according to JGZ guidelines}
\usage{
screen_curves(
  txt = "",
  loc = "",
  location = "",
  format = "1.0",
  legacy = TRUE,
  ...
)
}
\arguments{
\item{txt}{A JSON string, URL, or file with BDS data in JSON format. Data
should conform to the BDS JGZ 3.2.5 specification.}

\item{loc}{Deprecated. Use \code{session} instead.}

\item{location}{Legacy for \code{loc}}

\item{format}{JSON schema version, e.g., \code{"3.0"}. Used when uploading.}

\item{legacy}{Logical indicating whether legacy should be done.}

\item{...}{Ignored}
}
\value{
A JSON string containing a table with screening results
}
\description{
Screen growth curves according to JGZ guidelines
}
\note{
Deprecated for consistency. Function returns JSON, whereas all other
functions return the R object. The alternative \code{\link[=screen_growth]{screen_growth()}}
requests only results from screening. The alternative \code{\link[=custom_list]{custom_list()}}
produces the same list as \code{screen_curves}, but does not convert the
result to JSON.
}
\examples{
# # example json
# fn <- system.file("testdata", "client3.json", package = "james")
# fn <- system.file("testdata", "Laura_S_dev.json", package = "james")
#
# # first upload, then screen
# r1 <- jamesclient::james_post(path = "data/upload", txt = fn)
# location <- jamesclient::get_url(r1, "location")
# location
# screen_curves(loc = location)
#
# # upload & screen
# screen_curves(fn)
}
