% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_curves.R
\name{screen_curves}
\alias{screen_curves}
\title{Screen growth curves according to JGZ guidelines}
\usage{
screen_curves(
  txt = NULL,
  location = NULL,
  host = "https://groeidiagrammen.nl",
  path = "ocpu/library/james/R/convert_bds_ind",
  query = NULL
)
}
\arguments{
\item{txt}{a JSON string, URL or file}

\item{location}{A url that points to the server location where the
parsed data are stored. Optional. If specified, \code{location}
takes priority over \code{txt}. \code{location} is typically
the location of a previous \code{OpenCPU} call for uploading
the data.}

\item{host}{String with URL of the JAMES host machine. Defaults to
\code{https://groeidiagrammen.nl}.}

\item{path}{The end point of the request, typically an R function.
The default is \code{path = "ocpu/library/james/R/convert_bds_ind"}}

\item{query}{Passed down to \code{\link[httr]{modify_url}}.}
}
\value{
A JSON string
}
\description{
Screen growth curves according to JGZ guidelines
}
\examples{
\dontrun{
# example json
fn <- system.file("testdata", "client3.json", package = "james")

# first upload, then screen
r1 <- jamesclient::upload_bds(fn)
location <- jamesclient::get_url(r1, "location")
location
screen_curves(location = location)

# upload & screen
screen_curves(fn)
}
}
