% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{screen_growth-deprecated}
\alias{screen_growth-deprecated}
\alias{screen_growth}
\title{Screen growth curves according to JGZ guidelines}
\usage{
screen_growth(
  txt = "",
  loc = "",
  format = "1.0",
  ynames = c("hgt", "wgt", "hdc"),
  na.omit = TRUE,
  ...
)
}
\arguments{
\item{txt}{A JSON string, URL or file}

\item{loc}{Deprecated. Use \code{session} instead.}

\item{format}{String. JSON data schema version number. There are currently
three schemas supported: \code{"1.0"}, \code{"1.1"}, \code{"2.0"} and \code{"3.0"}.
Formats \code{"1.0"} and \code{"1.1"} are included for backward compatibility only.
Use \code{format = "3.0"} for new applications.}

\item{ynames}{Character vector identifying the measures to be
screened. By default, \code{ynames = c("hgt", "wgt", "hdc")}.}

\item{na.omit}{A logical indicating whether records with a missing
\code{x} (age) or \code{y} (yname) should be removed. Defaults to
\code{TRUE}.}

\item{...}{Ignored}
}
\description{
Screen growth curves according to JGZ guidelines
}
\examples{
host <- "http://localhost:8080"
fn <- system.file("testdata", "client3.json", package = "james")

\dontrun{
# first upload, then screen
r1 <- jamesclient::james_post(path = "data/upload", txt = fn)
location <- jamesclient::get_url(r1, "location")
location
screen_growth(loc = location)

session <- jamesclient::get_url(r1, "session")
session
screen_growth(session = session)

# upload & screen
screen_growth(fn)
}
}
