% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_growth.R
\name{screen_growth}
\alias{screen_growth}
\title{Screen growth curves according to JGZ guidelines}
\usage{
screen_growth(
  txt = "",
  loc = "",
  format = "1.0",
  ynames = c("hgt", "wgt", "hdc"),
  na.omit = TRUE,
  ...
)
}
\arguments{
\item{txt}{A JSON string, URL or file}

\item{loc}{Alternative to \code{txt}. Location where input data is uploaded
and converted to internal server format.}

\item{format}{String. JSON data schema version number. There are currently
three schemas supported: \code{"1.0"}, \code{"1.1"} and \code{"2.0"}. Formats \code{"1.0"} and
\code{"1.1"} are included for backward compatibility only. Use \code{format = "2.0"}
for new applications.}

\item{ynames}{Character vector identifying the measures to be
screened. By default, \code{ynames = c("hgt", "wgt", "hdc")}.}

\item{na.omit}{A logical indicating whether records with a missing
\code{x} (age) or \code{y} (yname) should be removed. Defaults to
\code{TRUE}.}

\item{...}{Ignored}
}
\description{
Screen growth curves according to JGZ guidelines
}
\note{
\code{screen_growth} superseeds \code{screen_curves} and will
only return results from growth screening.
}
\examples{
host <- "http://localhost"
fn <- system.file("testdata", "client3.json", package = "james")

\dontrun{
# first upload, then screen
r1 <- jamesclient::upload_txt(fn, host = host)
location <- jamesclient::get_url(r1, "location")
location
screen_growth(loc = location)

# upload & screen
screen_growth(fn)
}
}
