% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_chart.R
\name{select_chart}
\alias{select_chart}
\title{Selects the growth chart}
\usage{
select_chart(
  target = NULL,
  chartgrp = NULL,
  agegrp = NULL,
  sex = NULL,
  etn = NULL,
  ga = NULL,
  side = NULL,
  language = "dutch"
)
}
\arguments{
\item{target}{A list with elements \code{psn} (persondata) and \code{xyz} (timedata).}

\item{chartgrp}{The chart group: \code{'nl2010'}, \code{'preterm'}, \code{'who'}
or \code{character(0)}}

\item{agegrp}{Either \code{'0-15m'}, \code{'0-4y'}, \code{'1-21y'},
\code{'0-21y'} or \code{'0-4ya'}. Age group \code{'0-4ya'}
provides the 0-4 chart with weight for age (design E).}

\item{sex}{Either \code{'male'} or \code{'female'}}

\item{etn}{Either \code{'netherlands'}, \code{'turkish'},
\code{'moroccan'} or \code{'hindustani'}}

\item{ga}{Gestational age (in completed weeks)}

\item{side}{Either \code{'front'}, \code{'back'}, \code{'-hdc'} or
\code{'both'}}

\item{language}{Language: \code{'dutch'} or \code{'english'} (not
used)}
}
\value{
A list with elements \code{chartgrp}, \code{chartcode}
and \code{ga}
}
\description{
This function controls the behaviour for selecting a specific growth
chart based on a combination of individual data and user settings.
The default behaviour select preterm chart if gestational age is lower
or equal to 36 weeks, and determines the age group by the
maximum age found in the data.
}
\seealso{
\code{\link[chartcatalog:create_chartcode]{chartcatalog::create_chartcode()}}
}
