% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_data.R
\name{upload_data}
\alias{upload_data}
\title{Uploads, parses, converts and stores data on the server}
\usage{
upload_data(txt = "", ...)
}
\arguments{
\item{txt}{A JSON string, URL or file}

\item{...}{Passed down to \code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}}}
}
\value{
A tibble with a person attribute
}
\description{
Uploads JSON data that adhere to the BDS-format, parses its
contents, converts it to an internal JAMES target object,
and stores the result on the server for further processing.
The function is useful for caching input data over multiple requests to
\code{OpenCPU}. The cached data feed into other JAMES functions by means
of the \code{"loc"} argument. The server wipes the cached data after 2 hours.
}
\examples{
fn <- system.file("testdata", "client3.json", package = "james")
p <- upload_data(fn)
}
\seealso{
\code{\link[bdsreader:read_bds]{bdsreader::read_bds()}}
\code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}}
}
\author{
Stef van Buuren 2021
}
\keyword{server}
