% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_chartcode.R
\name{validate_chartcode}
\alias{validate_chartcode}
\title{Validates a growth chart code}
\usage{
validate_chartcode(chartcode = NULL)
}
\arguments{
\item{chartcode}{Character vector of chart codes,
typically something like \code{"NMAB"}}
}
\value{
Named logical vector with same length as \code{chartcode}.
A \code{TRUE} means a pass, a \code{FALSE} means a fail.
}
\description{
This function checks whether \code{chartcode} is a available
in the chart library.
}
\examples{
validate_chartcode(c("NJAA", "NZXX", "PMAAN26"))
}
\seealso{
\code{\link{list_charts}}
}
