% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{version}
\alias{version}
\title{Reports JAMES version}
\usage{
version(package = "james")
}
\arguments{
\item{package}{Package name}
}
\value{
A list with four elements:
\describe{
\item{package}{Package name}
\item{packageVersion}{Package version}
\item{packageDate}{Package date}
\item{Rversion}{R version}
}
}
\description{
This function return the version of the james package and R.
}
