% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate_jwt.R
\name{authenticate_jwt}
\alias{authenticate_jwt}
\title{Authentication request}
\usage{
authenticate_jwt(jwt = NULL, pubkey = NULL)
}
\arguments{
\item{jwt}{string containing the JSON Web Token (JWT)}

\item{pubkey}{path or object with RSA or EC public key, see \link[openssl:read_key]{openssl::read_pubkey}.}
}
\value{
A boolean
}
\description{
The function \code{authenticate_jwt} decrypts a JSON webtoken using a public key.
Returns \code{TRUE} if the request is granted to JAMES.
}
\examples{
key <- openssl::rsa_keygen(bits = 2048)
pubkey <- as.list(key)$pubkey
claim <- jose::jwt_claim(user = "test", session_key = 123, applications = "james;srm;psycat")
jwt <- jose::jwt_encode_sig(claim, key = key)
james:::authenticate_jwt(jwt, pubkey)
}
\author{
Arjan Huizing, Stef van Buuren 2021
}
\keyword{authentication}
