% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{authenticate}
\alias{authenticate}
\title{Authentication request}
\usage{
authenticate(authToken = NULL, pubkey = NULL, ...)
}
\arguments{
\item{authToken}{String containing the JSON Web Token (JWT)}

\item{pubkey}{Path or object with RSA or EC public key. When not given,
JAMES will search internally for the public key.}

\item{\dots}{Not used}
}
\value{
A boolean
}
\description{
The function \code{authenticate()} decrypts a JSON webtoken using a public key.
Returns \code{TRUE} if the request is granted to JAMES.
}
\examples{
key <- openssl::rsa_keygen(bits = 2048)
pubkey <- as.list(key)$pubkey
claim <- jose::jwt_claim(user = "test", session_key = 123, applications = "james;srm;psycat")
jwt <- jose::jwt_encode_sig(claim, key = key)
james:::authenticate(jwt, pubkey)
}
\author{
Arjan Huizing, Stef van Buuren 2021
}
\keyword{authentication}
