% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ddomain.R
\name{calculate_ddomain}
\alias{calculate_ddomain}
\title{Calculates the domain specific D-score and DAZ for each visit}
\usage{
calculate_ddomain(
  txt = "",
  session = "",
  format = "3.1",
  append = c("ddi", "gs1"),
  key = NULL,
  population = NULL,
  set = "GFCLS",
  domain = NULL,
  loc = "",
  ...
)
}
\arguments{
\item{txt}{A JSON string, URL, or file with BDS data in JSON format. Data
should conform to the BDS JGZ 3.2.5 specification.}

\item{session}{Optional session key if data is already uploaded to \code{sitehost}.}

\item{format}{JSON schema version, e.g., \code{"3.0"}. Used when uploading.}

\item{append}{Optional vector of strings indicating which instrument to base
D-score calculations on. Currently supports \code{ddi} and \code{gs1}. Requires JSON
schema V3.0 or later.}

\item{key}{String. They key identifies 1) the difficulty estimates
pertaining to a particular Rasch model, and 2) the prior mean and standard
deviation of the prior distribution for calculating the D-score.
The default key \code{NULL} sets \code{key = "gsed2510"}.
View \code{builtin_keys} for an overview of the available keys.}

\item{population}{String. The name of the reference population to calculate
DAZ.
Use \code{with(builtin_references, table(key, population))} to see which
built-in references are available for \code{key - population} combinations.
If not specified, the function set the default population as
\code{builtin_keys$base_population[key == builtin_keys$key]}.}

\item{set}{String. The name of the set of domains to use. See
\code{with(builtin_domaintable, table(set, domain))} for the domain names in each
set.}

\item{domain}{character vector of the name of the domain(s) for which to
compute the domain score. Per default all domains in the \code{set} are
used .}

\item{loc}{Deprecated. Use \code{session} instead.}

\item{...}{Ignored}
}
\value{
A list of data.frames.
}
\description{
The function \code{calculate_ddomain()} calculates five domain scores for
the D-score. Note that the domain scores may be highly variable due
to the small number of items per domain.
}
\examples{
fn <- system.file("testdata", "Laura_S.json", package = "james")
df <- calculate_dscore(txt = fn)
head(df, 4)
}
\author{
Iris Eekhout 2025
}
\keyword{server}
