% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dscore.R
\name{calculate_dscore}
\alias{calculate_dscore}
\title{Calculates the D-score and DAZ for each visit}
\usage{
calculate_dscore(
  txt = "",
  session = "",
  format = "1.0",
  output = c("table", "last_visit", "last_dscore"),
  loc = "",
  ...
)
}
\arguments{
\item{txt}{A JSON string, URL, or file with BDS data in JSON format. Data
should conform to the BDS JGZ 3.2.5 specification.}

\item{session}{Optional session key if data is already uploaded to \code{sitehost}.}

\item{format}{JSON schema version, e.g., \code{"3.0"}. Used when uploading.}

\item{output}{A string, either \code{"table"}, \code{"last_visit"} or
'\code{"last_dscore"} specifying the result. The default \code{"table"}
returns a table with four columns: \code{"date"}, \code{"x"} (age),
\code{"y"} (D-score) and \code{"z"} (DAZ). The number of rows equals to
the number of visits. If \code{output} equals \code{"last_visit"} the
function returns only the last row. If \code{output} equals
\code{"last_dscore"} the function returns only the D-score from the last row.}

\item{loc}{Deprecated. Use \code{session} instead.}

\item{...}{Ignored}
}
\value{
A table, row or scalar.
}
\description{
The function \code{draw_chart()} plots individual data on the growth chart.
}
\examples{
fn <- system.file("testdata", "Laura_S.json", package = "james")
df <- calculate_dscore(txt = fn)
head(df, 4)
}
\author{
Stef van Buuren 2020
}
\keyword{server}
