% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dscore.R
\name{calculate_dscore}
\alias{calculate_dscore}
\title{Calculates the D-score and DAZ for each visit}
\usage{
calculate_dscore(
  txt = "",
  loc = "",
  format = 2L,
  output = c("table", "last_visit", "last_dscore")
)
}
\arguments{
\item{txt}{A JSON string, URL or file with the data in JSON
format. The input data adhere to specification
\href{https://www.ncj.nl/themadossiers/informatisering/basisdataset/documentatie/?cat=12}{BDS JGZ 3.2.5},
and are converted to JSON according to \code{schema}.}

\item{loc}{Alternative to \code{txt}. Location where input data is uploaded
and converted to internal server format.}

\item{format}{Integer. JSON schema format number. There are currently two
schemas supported, formats \code{1} and \code{2}.}

\item{output}{A string, either \code{"table"}, \code{"last_visit"} or
'\code{"last_dscore"} specifying the result. The default \code{"table"}
returns a table with four columns: \code{"date"}, \code{"x"} (age),
\code{"y"} (D-score) and \code{"z"} (DAZ). The number of rows equals to
the number of visits. If \code{output} equals \code{"last_visit"} the
function returns only the last row. If \code{output} equals
\code{"last_dscore"} the function returns only the D-score from the last row.}
}
\value{
A table, row or scalar.
}
\description{
The function \code{draw_chart()} plots individual data on the growth chart.
}
\examples{
fn <- system.file("testdata", "Laura_S_dev.json", package = "james")
d <- calculate_dscore(txt = fn, format = 1)
}
\author{
Stef van Buuren 2020
}
\keyword{server}
