% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_bds_ind.R
\name{convert_bds_ind-deprecated}
\alias{convert_bds_ind-deprecated}
\alias{convert_bds_ind}
\title{Convert json BSD data for single individual to class individual}
\usage{
convert_bds_ind(txt = "", ...)
}
\arguments{
\item{txt}{A JSON string, URL or file}

\item{...}{Pass down to \code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}}}
}
\value{
A tibble with a person attribute.
}
\description{
This function takes data from a json source and saves as a tibble with
a person attribute.
}
\note{
Deprecated. Use \code{\link[=fetch_loc]{fetch_loc()}} instead.
}
\examples{
fn <- system.file("testdata", "client3.json", package = "james")
p <- convert_bds_ind(fn)
}
\author{
Stef van Buuren 2021
}
\keyword{server}
