% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{convert_bds_ind-deprecated}
\alias{convert_bds_ind-deprecated}
\alias{convert_bds_ind}
\title{Convert json BSD data for single individual to class individual}
\usage{
convert_bds_ind(txt = "", format = "1.0", ...)
}
\arguments{
\item{txt}{A JSON string, URL or file}

\item{format}{String. JSON data schema version number. There are currently
three schemas supported: \code{"1.0"}, \code{"1.1"}, \code{"2.0"}, \code{"3.0"} and \code{"3.1"}.
Formats \code{"1.0"} and \code{"1.1"} are included for backward compatibility only.
Use \code{format = "3.1"} for new applications.}

\item{...}{Ignored}
}
\value{
A list with elements \code{psn} (persondata) and \code{xyz} (timedata).
}
\description{
Convert json BSD data for single individual to class individual
}
\note{
Deprecated. Use \code{\link[=upload_data]{upload_data()}} instead.
}
\examples{
fn <- system.file("testdata", "client3.json", package = "james")
p <- convert_bds_ind(fn)
}
\author{
Stef van Buuren 2021
}
\keyword{server}
