% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_bds_ind.R
\name{convert_bds_ind}
\alias{convert_bds_ind}
\title{Convert json BSD data for single individual to class individual}
\usage{
convert_bds_ind(txt = NULL, ...)
}
\arguments{
\item{txt}{a JSON string, URL or file}

\item{...}{Additional parameter passed down to
\code{fromJSON(txt, ...)}, \code{new("xyz",... )} and
\code{new("bse",... )}. Useful parameters are \code{models =
"bsmodel"} for setting the broken stick model, or \code{call =
as.call(...)} for setting proper reference standards.}
}
\value{
An object of class \linkS4class{individual}.
}
\description{
This function takes data from a json source and saves as a an object
of class \linkS4class{individual}. The function automatically calculates
standard deviation scores and broken stick conditional means per visit.
}
\details{
This is function is a wrapper around \code{minihealth::convert_bds_individual()}.
}
\examples{
fn <- system.file("testdata", "client3.json", package = "james")
p <- convert_bds_ind(fn)
}
\seealso{
\linkS4class{individual}, \linkS4class{bse}, \linkS4class{xyz},
         \code{\link[minihealth]{convert_bds_individual}}
         \code{\link[jsonlite]{fromJSON}}
}
\author{
Stef van Buuren 2019
}
\keyword{server}
