% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ind_chartcodelist.R
\name{convert_ind_chartcodelist}
\alias{convert_ind_chartcodelist}
\title{Derive chartcode list from data}
\usage{
convert_ind_chartcodelist(ind_loc, chartcode, selector)
}
\arguments{
\item{ind_loc}{A url that points to the server location where the
data from a previous request to \code{convert_bds_ind()} are
stored. Optional. \code{ind_loc} takes priority over
\code{bds_data}.}

\item{chartcode}{The code of the requested growth chart, in the
case the \code{selector == "chartcode"}.}

\item{selector}{A string, either \code{"derive"}, \code{"data"} or
\code{"chartcode"}, that indicates the method to decide which
growth chart is drawn. Method \code{"derive"} (default)
calculates the chart from parameters \code{chartgrp},
\code{agegrp}, \code{sex}, \code{etn}, \code{ga} and \code{side}
parameters through the \code{select_chart()} function. Method
\code{"ind"} calculates the chart from the individual data.
Method \code{"chartcode"} will return the chart specified by the
\code{chartcode} parameter.}
}
\value{
A list with chart codes, produced by \code{chartbox::parse_chartcode}
}
\description{
The function load individual data that is already on the server,
calculates the chartcode and returns a list of parsed chartcode.
The function is called at initialization to automate the selection
of the chart according to the individual data.
}
\seealso{
\code{\link[chartbox]{parse_chartcode}}
}
\author{
Stef van Buuren 2019
}
\keyword{server}
