% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_tgt_chartadvice.R
\name{convert_tgt_chartadvice}
\alias{convert_tgt_chartadvice}
\title{Derive advice on chart choice from data}
\usage{
convert_tgt_chartadvice(
  txt = "",
  session = "",
  format = "1.0",
  chartcode = "",
  selector = c("data", "chartcode"),
  loc = "",
  ind_loc = "",
  ...
)
}
\arguments{
\item{txt}{A JSON string, URL, or file with BDS data in JSON format. Data
should conform to the BDS JGZ 3.2.5 specification.}

\item{session}{OpenCPU session key with the uploaded data}

\item{format}{JSON schema version, e.g., \code{"3.0"}. Used when uploading.}

\item{chartcode}{Optional. The code of the requested growth chart.}

\item{selector}{Either \code{"chartcode"}, \code{"data"} or \code{"derive"}.
The function can calculate the chart code by looking at the child
data (method \code{"data"}) or user input (method \code{"derive"}).
More in detail, the following behaviour decides between growth charts:
\describe{
\item{\code{"data"}}{Calculate chart code from the individual data.
This setting chooses the "optimal" chart for a given individual set of data.}
\item{\code{"derive"}}{Calculate chart code from a combination of user
data: \code{chartgrp}, \code{agegrp}, \code{side}, \code{sex},
\code{etn}, \code{ga}. The method does not use individual data. Use
this setting when chart choice needs to be reactive on user input.}
\item{\code{"chartcode"}}{Take string specified in \code{chartcode}}
}
If there is a valid \code{tgt} object, then the function simply obeys
the \code{selector} setting. If no valid \code{tgt} object is found,
the \code{"chartcode"} argument is taken. However, if the \code{"chartcode"}
is empty, then the function selects method \code{"derive"}.}

\item{loc}{Deprecated. Use \code{session} instead.}

\item{ind_loc}{Legacy. Will disappear in Nov 2022. Use \code{loc} instead.}

\item{...}{Ignored}
}
\value{
A list with the following elements
\describe{
\item{\code{population}}{A string identifying the population,
e.g. \code{'NL'},\code{'MA'}, \code{'TU'} or \code{'PT'}.}
\item{\code{sex}}{A string \code{"male"}, \code{"female"} or
\code{"undifferentiated"}.}
\item{\code{design}}{A letter indicating the chart design: \code{'A'} = 0-15m,
\code{'B'} = 0-4y, \code{'C'} = 1-21y, \code{'D'} = 0-21y, \code{'E'} =
0-4ya.}
\item{\code{side}}{A string indicating the side or \code{yname}:
\code{'front'}, \code{'back'}, \code{'both'}, \code{'hgt'},
\code{'wgt'}, \code{'hdc'}, \code{'bmi'}, \code{'wfh'}}
\item{\code{language}}{The language in which the chart is drawn. Currently only
\code{"dutch"} charts are implemented, but for \code{population == "PT"} we
may also have \code{"english"}.}
\item{\code{week}}{A scalar indicating the gestational age at birth.
Only used if \code{population == "PT"}.}
\item{\code{chartcode}}{A string indicating the chart code.}
\item{\code{chartgrp}}{A string indicating the chart group, either \code{"nl2010"},
\code{"preterm"}, \code{"who"}, \code{"gsed1"}, \code{"gsed1pt"}.}
\item{\code{agegrp}}{A string indicating the age group, either \code{"0-15m"},
\code{"0-4y"}, \code{"1-21y"} or \code{"0-21y"}.}
\item{\code{dnr}}{A string indicating the donor dataset for matching, either \code{"smocc"},
\code{"lollypop"}, \code{"terneuzen"} or \code{"pops"}.}
\item{\code{slider_list}}{A string indicating the set of slider labels, either \code{"0_2"},
\code{"0_4"} or \code{"0_29"}.}
\item{\code{period}}{A character vector of two elements, indicating the first and last period for the
matching analysis, e.g. like \code{c("3m", "14m")}.}
}
}
\description{
The function loads individual data from an url,
calculates the chartcode and returns a list of parsed chartcode
and age range of the data.
The function is called at initialization to automate setting
of proper chart and analysis defaults according to the child data.
}
\examples{
test25 <- system.file("extdata/bds_v3.0/test/test25.json", package = "jamesdemodata")
james:::convert_tgt_chartadvice(txt = test25)
}
\seealso{
\code{\link[chartcatalog:parse_chartcode]{chartcatalog::parse_chartcode()}}
}
\author{
Stef van Buuren 2020
}
\keyword{server}
