% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{custom_list}
\alias{custom_list}
\title{Provides a Screen growth curves according to JGZ guidelines}
\usage{
custom_list(txt = "", session = "", format = "1.0", loc = "", ...)
}
\arguments{
\item{txt}{A JSON string, URL, or file with BDS data in JSON format. Data
should conform to the BDS JGZ 3.2.5 specification.}

\item{session}{Optional session key if data is already uploaded to \code{sitehost}.}

\item{format}{JSON schema version, e.g., \code{"3.0"}. Used when uploading.}

\item{loc}{Deprecated. Use \code{session} instead.}

\item{...}{Ignored}
}
\value{
A table with screening results

A list with custom parts
}
\description{
Provides a Screen growth curves according to JGZ guidelines
}
\examples{
fn <- system.file("extdata", "bds_v2.0", "smocc", "Laura_S.json", package = "jamesdemodata")
host <- "http://localhost:8080"
\dontrun{
# first upload, then create custom list
r1 <- jamesclient::james_post(host = host, path = "data/upload", txt = fn)
loc <- jamesclient::get_url(r1, "location")
list1 <- custom_list(loc = loc)

# upload & screen
list2 <- custom_list(txt = fn)
identical(list1, list2)
}
}
