% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_list.R
\name{custom_list}
\alias{custom_list}
\title{Provides a Screen growth curves according to JGZ guidelines}
\usage{
custom_list(txt = "", loc = "", version = 1L)
}
\arguments{
\item{txt}{A JSON string, URL or file with the data in JSON
format. The input data adhere to specification
\href{https://www.ncj.nl/themadossiers/informatisering/basisdataset/documentatie/?cat=12}{BDS JGZ 3.2.5},
and are converted to JSON according to \code{schema}.}

\item{loc}{Alternative to \code{txt}. Location where input data is uploaded
and converted to internal server format.}

\item{version}{Integer. JSON schema version number. There are currently two
schemas supported, versions \code{1} and \code{2}.}
}
\value{
A table with screening results

A list with custom parts
}
\description{
Provides a Screen growth curves according to JGZ guidelines
}
\examples{
fn <- system.file("extdata", "bds_str", "smocc", "Laura_S.json", package = "jamesdemodata")
host <- "http://localhost"
\dontrun{
# first upload, then create custom list
r1 <- jamesclient::upload_txt(fn, host = host)
loc <- jamesclient::get_url(r1, "location")
list1 <- custom_list(loc = loc)

# upload & screen
list2 <- custom_list(txt = fn)
identical(list1, list2)
}
}
