% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcat_next.R
\name{dcat_next}
\alias{dcat_next}
\title{Calculates the next item to administer in an adaptive test to measure the
D-score. Uses the previously administered items and item scores as input.}
\usage{
dcat_next(
  txt = "",
  instrument = "gs1",
  key = "gsed2510",
  p = 50,
  session = "",
  format = "3.1",
  ...
)
}
\arguments{
\item{txt}{A JSON string, URL, or file with BDS data in JSON format. Data
should conform to the BDS JGZ 3.2.5 specification.}

\item{instrument}{A character vector with 3-position codes of instruments
that should match. The default is \code{instrument = "gs1"} for GSED SF;
\code{instrument = NULL} allows for all instruments.}

\item{key}{String. They key identifies 1) the difficulty estimates
pertaining to a particular Rasch model, and 2) the prior mean and standard
deviation of the prior distribution for calculating the D-score.
The default key \code{key = "gsed2510"}.}

\item{p}{percentage to pass the item, difficulty in percentile units.}

\item{session}{Optional session key if data is already uploaded to \code{sitehost}.}

\item{format}{JSON schema version, e.g., \code{"3.0"}. Used when uploading.}

\item{...}{Ignored}
}
\value{
A string of milestones
}
\description{
Calculates the next item to administer in an adaptive test to measure the
D-score. Uses the previously administered items and item scores as input.
}
\examples{
txt <- system.file("examples", "example_v3.1.json", package = "bdsreader")
dcat_next(txt = txt, p = 50)
}
\author{
Iris Eekhout 2025
}
