% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcat_start.R
\name{dcat_start}
\alias{dcat_start}
\title{Calculates the start item from age for an adaptive test to measure D-score.}
\usage{
dcat_start(
  txt = "",
  instrument = "gs1",
  key = "gsed2510",
  population = "",
  p = 50,
  session = "",
  format = "3.0",
  loc = "",
  ...
)
}
\arguments{
\item{txt}{A JSON string, URL, or file with BDS data in JSON format. Data
should conform to the BDS JGZ 3.2.5 specification.}

\item{instrument}{A character vector with 3-position codes of instruments
that should match. The default is \code{instrument = gs1} for GSED SF;
\code{instrument = NULL} allows for all instruments.}

\item{key}{String. They key identifies 1) the difficulty estimates
pertaining to a particular Rasch model, and 2) the prior mean and standard
deviation of the prior distribution for calculating the D-score.
The default key \code{key = "gsed2510"}.}

\item{population}{String. Name of the reference population.}

\item{p}{percentage to pass the item, difficulty in percentile units.}

\item{session}{Optional session key if data is already uploaded to \code{sitehost}.}

\item{format}{JSON schema version, e.g., \code{"3.0"}. Used when uploading.}

\item{loc}{Deprecated. Use \code{session} instead.}

\item{...}{Ignored}
}
\value{
A string of milestones.
}
\description{
Calculates the start item from age for an adaptive test to measure D-score.
}
\examples{
fn <- system.file("examples", "example_v3.1.json", package = "bdsreader")
dcat_start(txt = fn, p = 50)
}
\author{
Iris Eekhout 2025
}
