% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcat_start.R
\name{dcat_start}
\alias{dcat_start}
\title{Calculates the start item from age for an adaptive test to measure D-score.}
\usage{
dcat_start(
  txt = "",
  instrument = "gs1",
  key = "gsed2510",
  population = NULL,
  p = 50,
  session = "",
  format = "3.1",
  ...
)
}
\arguments{
\item{txt}{A JSON string, URL, or file with BDS data in JSON format. Data
should conform to the BDS JGZ 3.2.5 specification.}

\item{instrument}{A character vector with 3-position codes of instruments
that should match. The default is \code{instrument = "gs1"} for GSED SF;
\code{instrument = NULL} allows for all instruments.}

\item{key}{String. They key identifies 1) the difficulty estimates
pertaining to a particular Rasch model, and 2) the prior mean and standard
deviation of the prior distribution for calculating the D-score.
The default key \code{NULL} sets \code{key = "gsed2406"}.
View \code{builtin_keys} for an overview of the available keys.}

\item{population}{String. The name of the reference population to calculate
DAZ.
Use \code{with(builtin_references, table(key, population))} to see which
built-in references are available for \code{key - population} combinations.
If not specified, the function set the default population as
\code{builtin_keys$base_population[key == builtin_keys$key]}.}

\item{p}{percentage to pass the item, difficulty in percentile units.}

\item{session}{Optional session key if data is already uploaded to \code{sitehost}.}

\item{format}{JSON schema version, e.g., \code{"3.0"}. Used when uploading.}

\item{...}{Ignored}
}
\value{
A string of milestones.
}
\description{
Calculates the start item from age for an adaptive test to measure D-score.
}
\examples{
fn <- system.file("examples", "example_v3.1.json", package = "bdsreader")
dcat_start(txt = fn, p = 50)
}
\author{
Iris Eekhout 2025
}
