% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_chart_bds.R, R/draw_chart_ind.R
\name{draw_chart_bds}
\alias{draw_chart_bds}
\alias{draw_chart_ind}
\title{Convert bds-format data to individual and return growth chart}
\usage{
draw_chart_bds(txt = NULL, chartcode = NULL,
  curve_interpolation = TRUE, ...)

draw_chart_ind(location = NULL, chartcode = NULL,
  curve_interpolation = TRUE, ...)
}
\arguments{
\item{txt}{A JSON string, URL or file}

\item{chartcode}{The code of the requested growth chart. If not
specified, the server will automatically plot child height for
the most recent age period.}

\item{curve_interpolation}{A logical indicating whether curve interpolation shoud be applied.}

\item{\dots}{For \code{draw_chart_bds}, additional parameter passed
down to \code{fromJSON(txt, ...)}, \code{new("xyz",... )} and
\code{new("bse",... )}. Useful parameters are \code{models =
"bsmodel"} for setting the broken stick model, or \code{call =
as.call(...)} for setting proper reference standards.}

\item{location}{A url that points to the server location where the
data from a previous request to \code{convert_bds_ind()} are stored.}
}
\value{
tbd
}
\description{
The function \code{draw_chart_bds()} convert bds data into an object
of class \linkS4class{individual}, and then draws the
individual data on the requested growth chart.

The function \code{draw_chart_ind()} expect an input location from
a previous call, and plots the individual data on the requested
growth chart.
}
\examples{
fn <- file.path(path.package("james"), "testdata", "client3.json")
g <- draw_chart_bds(txt = fn)
}
\seealso{
\linkS4class{individual},
\code{\link[groeidiagrammen]{select_chart}}
\code{\link[groeidiagrammen]{draw_plot}}
}
\author{
Stef van Buuren 2019
}
\keyword{server}
