% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_chart.R
\name{draw_chart}
\alias{draw_chart}
\title{Draw growth chart}
\usage{
draw_chart(
  txt = NULL,
  loc = NULL,
  chartcode = NULL,
  selector = c("data", "derive"),
  curve_interpolation = TRUE,
  dnr = c("smocc", "terneuzen", "lollypop.preterm", "lollypop.term", "pops"),
  lo = NULL,
  hi = NULL,
  nmatch = 0L,
  exact_sex = TRUE,
  exact_ga = FALSE,
  break_ties = FALSE,
  show_realized = FALSE,
  show_future = FALSE,
  ...
)
}
\arguments{
\item{txt}{A JSON string, URL or file, with the data (in JSON
format) to be uploaded. The variable specification are expected
to be according specification
\href{https://www.ncj.nl/themadossiers/informatisering/basisdataset/documentatie/?cat=12}{BDS
 JGZ 3.2.5}, and converted to JSON.}

\item{loc}{Alternative to \code{txt}. Location where data is uploaded
and processed, for example, as obtained by \code{upload_txt()}.}

\item{chartcode}{Optional. The code of the requested growth chart.
If not specified or if not valid, the function calculates the chart code
by the method specified by the \code{selector} argument.}

\item{selector}{Either \code{data} or \code{"derive"} specifying the
method to decide between growth charts:
  \describe{
  \item{\code{"data"}}{Calculate chart code from the individual data by
  \code{select_chart()} (default).}
  \item{\code{"derive"}}{Calculate chart code from parameters
  \code{chartgrp}, \code{agegrp}, \code{sex}, \code{etn}, \code{ga}
  and \code{side} by \code{select_chart()}.}
  \item{\code{"chartcode"}}{Take string specified in \code{chartcode}.}
  }}

\item{curve_interpolation}{A logical indicating whether curve
interpolation shoud be applied.}

\item{dnr}{A string with the name of the donor data
(currently available are \code{smocc}, \code{terneuzen},
\code{lollypop.preterm} or \code{lollypop.term})}

\item{lo}{Value of the left visit coded as string, e.g. \code{"4w"}
or \code{"7.5m"}}

\item{hi}{Value of the right visit coded as string, e.g. \code{"4w"}
or \code{"7.5m"}}

\item{nmatch}{Integer. Number of matches needed. When
\code{nmatch == 0L} no matches are sought.}

\item{exact_sex}{A logical indicating whether sex should be
matched exactly}

\item{exact_ga}{A logical indicating whether gestational age
should be matched exactly}

\item{break_ties}{A logical indicating whether ties should broken
randomly. The default (\code{TRUE}) breaks ties randomly.}

\item{show_realized}{A logical indicating whether the realized
growth of the target child should be drawn}

\item{show_future}{A logical indicating whether the predicted
growth of the target child should be drawn}

\item{\dots}{Parameter passed down to the \code{select_chart()} function,
such as \code{chartgrp}, \code{agegrp}, \code{sex}, \code{etn},
\code{ga} and \code{side}.}
}
\value{
A \code{gTree} object.
}
\description{
The function \code{draw_chart()} plots individual data on the growth chart.
}
\examples{
fn <- system.file("testdata", "client3.json", package = "james")
g <- draw_chart(txt = fn)
}
\seealso{
\linkS4class{individual}, \code{\link{select_chart}}
}
\author{
Stef van Buuren 2020
}
\keyword{server}
