% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_chart.R, R/draw_chart_bds.R,
%   R/draw_chart_ind.R
\name{draw_chart}
\alias{draw_chart}
\alias{draw_chart_bds}
\alias{draw_chart_ind}
\title{Draw growth chart}
\usage{
draw_chart(
  txt = NULL,
  loc = NULL,
  selector = c("derive", "data", "chartcode"),
  chartcode = NULL,
  chartgrp = NULL,
  agegrp = NULL,
  sex = NULL,
  etn = NULL,
  ga = NULL,
  side = "hgt",
  curve_interpolation = TRUE,
  dnr = c("smocc", "terneuzen", "lollypop.preterm", "lollypop.term", "pops"),
  lo = NULL,
  hi = NULL,
  nmatch = 0L,
  exact_sex = TRUE,
  exact_ga = FALSE,
  break_ties = FALSE,
  show_realized = FALSE,
  show_future = FALSE,
  ...
)

draw_chart_bds(txt = NULL, chartcode = NULL, curve_interpolation = TRUE, ...)

draw_chart_ind(loc = NULL, chartcode = NULL, curve_interpolation = TRUE, ...)
}
\arguments{
\item{txt}{A JSON string, URL or file}

\item{loc}{A url that points to the server location where the
data from a previous request to \code{convert_bds_ind()} are
stored. Optional. \code{loc} takes priority over
\code{txt}.}

\item{selector}{A string, either \code{"derive"}, \code{"data"} or
\code{"chartcode"}, that indicates the method to decide which
growth chart is drawn. Method \code{"derive"} (default)
calculates the chart from parameters \code{chartgrp},
\code{agegrp}, \code{sex}, \code{etn}, \code{ga} and \code{side}
parameters through the \code{select_chart()} function. Method
\code{"data"} calculates the chart from the individual data.
Method \code{"chartcode"} will return the chart specified by the
\code{chartcode} parameter.}

\item{chartcode}{The code of the requested growth chart, in the
case the \code{selector == "chartcode"}.}

\item{chartgrp}{The chart group: \code{'nl2010'}, \code{'preterm'}, \code{'who'}
or \code{character(0)}}

\item{agegrp}{Either \code{'0-15m'}, \code{'0-4y'}, \code{'1-21y'},
\code{'0-21y'} or \code{'0-4ya'}. Age group \code{'0-4ya'}
provides the 0-4 chart with weight for age (design E).}

\item{sex}{Either \code{'male'} or \code{'female'}}

\item{etn}{Either \code{'netherlands'}, \code{'turkish'},
\code{'moroccan'} or \code{'hindustani'}}

\item{ga}{Gestational age (in completed weeks)}

\item{side}{Either \code{'front'}, \code{'back'}, \code{'-hdc'} or
\code{'both'}}

\item{curve_interpolation}{A logical indicating whether curve
interpolation shoud be applied.}

\item{dnr}{A string with the name of the donor data
(currently available are \code{smocc}, \code{terneuzen},
\code{lollypop.preterm} or \code{lollypop.term})}

\item{lo}{Value of the left visit coded as string, e.g. \code{"4w"}
or \code{"7.5m"}}

\item{hi}{Value of the right visit coded as string, e.g. \code{"4w"}
or \code{"7.5m"}}

\item{nmatch}{Integer. Number of matches needed. When
\code{nmatch == 0L} no matches are sought.}

\item{exact_sex}{A logical indicating whether sex should be
matched exactly}

\item{exact_ga}{A logical indicating whether gestational age
should be matched exactly}

\item{break_ties}{A logical indicating whether ties should broken
randomly. The default (\code{TRUE}) breaks ties randomly.}

\item{show_realized}{A logical indicating whether the realized
growth of the target child should be drawn}

\item{show_future}{A logical indicating whether the predicted
growth of the target child should be drawn}

\item{\dots}{For \code{draw_chart_bds}, additional parameter passed
down to \code{fromJSON(txt, ...)}, \code{new("xyz",... )} and
\code{new("bse",... )}. Useful parameters are \code{models =
"bsmodel"} for setting the broken stick model, or \code{call =
as.call(...)} for setting proper reference standards.}
}
\value{
tbd
}
\description{
The function \code{draw_chart()} can read data from an input
location from a previous call, calculate the chartcode, and plot
the individual data on the requested growth chart.

The function \code{draw_chart_bds()} convert bds data into an object
of class \linkS4class{individual}, and then draws the
individual data on the requested growth chart.

The function \code{draw_chart_ind()} expect an input location from
a previous call, and plots the individual data on the requested
growth chart.
}
\examples{
fn <- system.file("testdata", "client3.json", package = "james")
g <- draw_chart_bds(txt = fn)
}
\seealso{
\linkS4class{individual}, \code{\link{select_chart}}

\linkS4class{individual},
\code{\link{select_chart}}
\code{\link[chartplotter]{process_chart}}
}
\author{
Stef van Buuren 2019
}
\keyword{server}
