% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_bds_chart.R
\name{draw_chart_bds}
\alias{draw_chart_bds}
\title{Convert bds-format data to individual and return growth chart}
\usage{
draw_chart_bds(txt = NULL, chartcode = NULL,
  curve_interpolation = TRUE, ...)
}
\arguments{
\item{txt}{A JSON string, URL or file}

\item{chartcode}{The code of the requested growth chart. If not
specified, the server will automatically plot child height for
the most recent age period.}

\item{curve_interpolation}{A logical indicating whether curve interpolation shoud be applied.}

\item{\dots}{Additional parameter passed down to
\code{fromJSON(txt, ...)}, \code{new("xyz",... )} and
\code{new("bse",... )}. Useful parameters are \code{models =
"bsmodel"} for setting the broken stick model, or \code{call =
as.call(...)} for setting proper reference standards.}
}
\value{
-specify-
}
\description{
This function takes data from a json source and saves as a an object
of class \linkS4class{individual}. The function then draws the
individual data on the requested growth chart.
}
\examples{
fn <- file.path(path.package("james"), "testdata", "client3.json")
g <- draw_chart_bds(txt = fn)
}
\seealso{
\linkS4class{individual},
\code{\link[groeidiagrammen]{select_chart}}
\code{\link[groeidiagrammen]{draw_plot}}
}
\author{
Stef van Buuren 2019
}
\keyword{server}
