% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_plot.R
\name{draw_plot}
\alias{draw_plot}
\title{Constructs and plots the A4 growth chart}
\usage{
draw_plot(ind, chartcode, curve_interpolation = TRUE, quiet = TRUE)
}
\arguments{
\item{ind}{An S4 object of class \code{individual} containing data of the individual}

\item{chartcode}{A string with chart code}

\item{curve_interpolation}{A logical indicating whether curve interpolation shoud be applied.}

\item{quiet}{Logical indicating whether chart code should be written
to standard output. Default is \code{quiet = TRUE}.}
}
\value{
A grid object
}
\description{
This function is a wrapper of \code{groeidiagrammen::draw_plot}
}
