% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_loc.R
\name{fetch_loc}
\alias{fetch_loc}
\title{Uploads, parses, converts and stores data on the server for further processing}
\usage{
fetch_loc(txt = "", format = "1.0", ...)
}
\arguments{
\item{txt}{A JSON string, URL or file}

\item{format}{String. JSON data schema version number. There are currently
three schemas supported: \code{1.0}, \code{1.1} and \code{2.0}.}

\item{...}{Passed down to \code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}}}
}
\value{
A tibble with a person attribute
}
\description{
Uploads JSON data that adhere to the BDS-format, parses its
contents, converts it to an internal JAMES target object,
and stores the result on the server for further processing.
The function is useful for caching input data over multiple requests to
\code{OpenCPU}. The cached data feed into other JAMES functions by means
of the \code{"loc"} argument. The server wipes the cached data after 24 hours.
}
\examples{
fn <- system.file("testdata", "client3.json", package = "james")
p <- fetch_loc(fn)
}
\seealso{
\code{\link[bdsreader:read_bds]{bdsreader::read_bds()}}
\code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}}
}
\author{
Stef van Buuren 2021
}
\keyword{server}
