% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get_session_object}
\alias{get_session_object}
\title{Load data from a previous OpenCPU session on same host}
\usage{
get_session_object(session, object = ".val")
}
\arguments{
\item{session}{Character, e.g. \code{session = "x077dd78bd0bbc6"}}

\item{object}{Character, e.g. \code{object = ".val"}. Refers to objects within the
\verb{/R} path (e.g. function returns or variables saved in scripts).}
}
\value{
If found, object from a the session. If not found, the function generates a
warning and return \code{NULL}.
}
\description{
This function loads data from a previous OpenCPU session on the same host.
The assumption is that sessions are stored in path \verb{/tmp/ocpu-store}.
Provides a short-cut to the data eliminating the need to make self-requests.
}
\examples{
\dontrun{
get_session_object(session = "x02a93ec661121", object = ".val")
}
}
