% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j_get.R
\name{j_get}
\alias{j_get}
\title{Get data from archive}
\usage{
j_get(index, type, version, scenario = james.env$j_root$active_scenario,
  project = james.env$j_root$active_project, what = c("data", "fig",
  "object"))
}
\arguments{
\item{index}{unique id for your data}

\item{type}{data type you want to get}

\item{version}{version you want to get (default is active (c.q. newest) version)}

\item{scenario}{scenario from which you want to get the data (default is active scenario)}

\item{project}{project from which you want to get the data (default is active project)}

\item{what}{data type you want to get is one from c("data", "fig", "object") (default "data")}
}
\value{
data, JFig or JData object (NULL if not present)
}
\description{
Get data from archive
}
\examples{
j_put(x = 4:12, type = "age", scenario = "Primary school", project = "Education")
j_get(type = "age")
j_put(x = 3:13, type = "age")
j_ls()
j_ls(collapsed = FALSE)
j_get(index = 1)
j_get(type = "age")
j_get(type = "age", version = 1, scenario = "Primary school", project = "Education")

}
\seealso{
\code{\link{j_put}}, \code{\link{j_ls}}, \code{\link{j_init}}
}
