% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j_init.R
\name{j_init}
\alias{j_init}
\title{Initialize James}
\usage{
j_init(file_name = paste0(Sys.getenv("HOME"), .Platform$file.sep,
  Sys.info()["user"], ".james"), active_scenario, active_project)
}
\arguments{
\item{file_name}{File holding your archive. Defaults to <username>.james in your home directory, where <username> is your username}

\item{active_scenario}{Scenario you want to work on}

\item{active_project}{Project you want to work on}
}
\value{
JRoot object holding your data. Please assign return value to .j_root, as James assumes all data are globally available in .j_root. See examples.
}
\description{
Creates or opens a file and a JRoot object holding your data
}
\examples{
.j_root <- j_init()
j_ls()

}
\seealso{
\code{\link{j_put}}, \code{\link{j_get}}, \code{\link{j_ls}}
}
