% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_charts.R
\name{list_charts}
\alias{list_charts}
\title{List the available growth charts}
\usage{
list_charts(authToken = NULL, ...)
}
\arguments{
\item{authToken}{A JSON web token.}

\item{\dots}{Used for authentication}
}
\description{
List the available growth charts
}
