% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_screeners.R
\name{list_screeners}
\alias{list_screeners}
\title{List the available growth screeners}
\usage{
list_screeners(...)
}
\arguments{
\item{\dots}{Used for authentication, passed down to \code{\link[growthscreener:list_screeners]{growthscreener::list_screeners()}}.}
}
\description{
List the available growth screeners
}
\examples{
head(list_screeners(ynames = "hgt"), 2)
}
