% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview_proeftuin.R
\name{preview_screeners}
\alias{preview_screeners}
\title{Preview growth screener results}
\usage{
preview_screeners(txt = "", session = "", ...)
}
\arguments{
\item{txt}{A JSON string, URL or file}

\item{session}{Optional session key if data is already uploaded to \code{sitehost}.}

\item{...}{Ignored}
}
\value{
A data frame.
}
\description{
Experimental "Proeftuin" endpoint: thin wrapper around
\code{\link[=apply_screeners]{apply_screeners()}} for the Proeftuin preview UI.
}
\examples{
fn <- system.file("testdata", "client3.json", package = "james")
preview_screeners(fn)
}
\seealso{
\code{\link[=apply_screeners]{apply_screeners()}}
}
\author{
Stef van Buuren 2026
}
\keyword{server}
