% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_blend.R
\name{request_blend}
\alias{request_blend}
\title{Provides multiple outputs in one request}
\usage{
request_blend(
  txt = "",
  sitehost = "",
  session = "",
  blend = "standard",
  loc = "",
  ...
)
}
\arguments{
\item{txt}{A JSON string, URL or file with the data in JSON
format. The input data adhere to specification
\href{https://www.ncj.nl/themadossiers/informatisering/basisdataset/documentatie/?cat=12}{BDS JGZ 3.2.5},
and are converted to JSON according to \code{schema}.}

\item{sitehost}{The host that renders the site. Normally, that would be equal
to the host on which JAMES runs. If not specified, the function throws a warning
and sets \code{sitehost} to \code{"http://localhost"}.}

\item{session}{Alternative to \code{txt}. Session key where input data is uploaded
on \code{sitehost}.}

\item{blend}{A string indicating the requested blend. The default (\code{"standard"})
returns the results of the standard end points that produces tables. Graphs
are currently not supported.}

\item{loc}{Alternative to \code{txt}. Location where input data is uploaded.
Argument \code{loc} is deprecated and will disappear in Sept 2022; please
use \code{session} instead.}

\item{...}{Ignored}
}
\value{
The default \code{blend = "standard"} return a list with the following
components:
\describe{
\item{\code{txt}}{String, file or URL with child data}
\item{\code{session}}{Session with uploaded child data}
\item{\code{child}}{Processed child level data}
\item{\code{time}}{Processed time level data}
\item{\code{screeners}}{Results from application of screeners to child data}
\item{\code{site}}{URL with personalised child data}
}
}
\description{
Function \code{request_blend()} acts as a one-stop-shop to obtain multiple
outputs through one request.
}
\examples{
\dontrun{
fn <- system.file("extdata", "bds_v2.0", "smocc", "Laura_S.json", package = "jamesdemodata")
results <- request_blend(txt = fn)
}
}
