% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_blend.R
\name{request_blend}
\alias{request_blend}
\title{Provides a Screen growth curves according to JGZ guidelines}
\usage{
request_blend(txt = "", loc = "", blend = "standard", ...)
}
\arguments{
\item{txt}{A JSON string, URL or file with the data in JSON
format. The input data adhere to specification
\href{https://www.ncj.nl/themadossiers/informatisering/basisdataset/documentatie/?cat=12}{BDS JGZ 3.2.5},
and are converted to JSON according to \code{schema}.}

\item{loc}{Alternative to \code{txt}. Location where input data is uploaded
and converted to internal server format.}

\item{blend}{A string indicating the requested blend. The default (\code{"standard"})
returns the results of all standard end points.}

\item{...}{Ignored}
}
\value{
The default \code{blend = "standard"} return a list with the following
components:
\describe{
\item{\code{txt}}{String, file or URL with child data}
\item{\code{loc}}{URL of uploaded and processed child data}
\item{\code{child}}{Processed child level data}
\item{\code{time}}{Processed time level data}
\item{\code{chart}}{Growth chart figure (svglite)}
\item{\code{screeners}}{Results from application of screeners to child data}
\item{\code{site}}{URL with personalised child data}
}
}
\description{
Provides a Screen growth curves according to JGZ guidelines
}
\examples{
fn <- system.file("extdata", "bds_v2.0", "smocc", "Laura_S.json", package = "jamesdemodata")
results <- request_blend(txt = fn)
}
