% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_blend.R
\name{request_blend}
\alias{request_blend}
\title{Provides a Screen growth curves according to JGZ guidelines}
\usage{
request_blend(txt = "", loc = "", blend = "standard", ...)
}
\arguments{
\item{txt}{A JSON string, URL or file with the data in JSON
format. The input data adhere to specification
\href{https://www.ncj.nl/themadossiers/informatisering/basisdataset/documentatie/?cat=12}{BDS JGZ 3.2.5},
and are converted to JSON according to \code{schema}.}

\item{loc}{Alternative to \code{txt}. Location where input data is uploaded
and converted to internal server format.}

\item{blend}{A string indicating the requested blend. The default (\code{"standard"})
returns all standard end points.}

\item{...}{Ignored}
}
\value{
A table with screening results

A list with custom parts
}
\description{
Provides a Screen growth curves according to JGZ guidelines
}
\examples{
fn <- system.file("extdata", "bds_v2.0", "smocc", "Laura_S.json", package = "jamesdemodata")
results <- request_blend(txt = fn)
\dontrun{
# the more complicated way
# first upload, then create custom list
host <- "http://localhost"
r1 <- jamesclient::upload_txt(fn, host = host)
loc <- jamesclient::get_url(r1, "location")
list1 <- request_blend(loc = loc)
}
}
