% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_site.R
\name{request_site}
\alias{request_site}
\title{Request site containing personalised charts}
\usage{
request_site(txt = NULL, loc = NULL, schema = NULL)
}
\arguments{
\item{txt}{A JSON string, URL or file, with the data (in JSON
format) to be uploaded. The variable specification are expected
to be according specification
\href{https://www.ncj.nl/themadossiers/informatisering/basisdataset/documentatie/?cat=12}{BDS
 JGZ 3.2.5}, and converted to JSON.}

\item{loc}{Alternative to \code{txt}. Location where data is uploaded
and processed, for example, as obtained by \code{upload_txt()}.}

\item{schema}{Optional. A JSON string, URL or file that selects the JSON validation
schema. Only relevant if \code{txt} is specified.}
}
\value{
URL composed of JAMES server and query string starting
with \code{?ind=...}, which indicates the URL of the uploaded
child data
}
\description{
Server-side function cto construct a URL to a site that shows a personalised
growth chart. The site includes a navigation bar so that the end
user can tweak chart choice.
}
\details{
The function can take child data in two forms: 1) a JSON file with
BDS-formatted child data (argument \code{txt}) or, 2) a URL with
child data on a previously stored server location on the server
(argument \code{loc}).

One of \code{txt} or \code{loc} need to be specified. If both are given,
\code{txt} takes precedence. The \code{txt} option uploads the data to the
host that runs \code{request_side()}.
}
\note{
This not yet work with Docker because we cannot get the host URL.
}
\examples{
fn <- system.file("extdata", "smocc", "Laura_S.json", package = "jamestest")

# as two steps (upload to default site groeidiagrammen.nl)
resp <- upload_txt(fn)
loc <- jamesclient::get_url(resp, "location")
request_site(loc = loc)

\dontrun{
# upload to localhost (if running)
resp <- upload_txt(fn, "http://localhost")
loc <- jamesclient::get_url(resp, "location")
request_site(loc = loc)

# as one step
request_site(fn)
}
}
\seealso{
\code{\link{upload_txt}}, \code{\link[jamesclient]{get_url}}
}
