% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_site.R
\name{request_site}
\alias{request_site}
\title{Request site containing personalised charts}
\usage{
request_site(txt = "", loc = "", schema = NULL, upload = FALSE, host = NULL)
}
\arguments{
\item{txt}{A JSON string, URL or file with the data in JSON
format. The input data adhere to specification
\href{https://www.ncj.nl/themadossiers/informatisering/basisdataset/documentatie/?cat=12}{BDS JGZ 3.2.5},
and are converted to JSON according to \code{schema}.}

\item{loc}{Alternative to \code{txt}. Location where input data is uploaded
and converted to internal server format.}

\item{schema}{Optional. A JSON string, URL or file that selects the JSON validation
schema. Only used if the \code{txt} argument is specified.}

\item{upload}{Logical. If \code{TRUE} then \code{request_site()} will upload
the data in \code{txt} and return a site address with the \code{?loc=} query appended.
The default (\code{FALSE}) just appends \code{?txt=} to the site url, thus
deferring validation and conversion to internal representation to the site.}

\item{host}{URL of the JAMES server. By default, host is the currently
running server that processes the request.}
}
\value{
URL composed of JAMES server, possibly appended by query string starting
with \code{?txt=} or \code{?loc=}.
}
\description{
Server-side function to construct a URL to a site that shows a personalised
growth chart. The site includes a navigation bar so that the end
user can interact to chart choices.
}
\details{
One of \code{txt} or \code{loc} needs to be specified. If both are given,
\code{txt} takes precedence. If neither is given, then the function returns
the base site without any data.
}
\note{
This function does not yet work with Docker because we cannot get the
host URL name.
The output form \code{?txt=} currently ignores the \code{schema} argument.
}
\examples{
fn <- system.file("testdata", "client3.json", package = "james")
js <- jsonlite::toJSON(jsonlite::fromJSON(fn), auto_unbox = TRUE)
url <- "https://groeidiagrammen.nl/ocpu/library/james/testdata/client3.json"

# solutions that create a ?loc= parameter

# upload JSON string to default host, return site url
site <- request_site(js, upload = TRUE)
site
browseURL(site)
request_site(url, upload = TRUE)

# same, as two steps, starting from file name
resp <- upload_txt(fn)
loc <- jamesclient::get_url(resp, "location")
request_site(loc = loc)

# solutions that create a ?txt= parameter

request_site(fn)
request_site(js)
request_site(url)
}
\seealso{
\code{\link{upload_txt}}, \code{\link[jamesclient]{get_url}}
}
