% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_curves.R
\name{screen_curves-deprecated}
\alias{screen_curves-deprecated}
\alias{screen_curves}
\title{Screen growth curves according to JGZ guidelines}
\usage{
screen_curves(txt = "", loc = "", location = "", version = 1L, legacy = TRUE)
}
\arguments{
\item{txt}{A JSON string, URL or file with the data in JSON
format. The input data adhere to specification
\href{https://www.ncj.nl/themadossiers/informatisering/basisdataset/documentatie/?cat=12}{BDS JGZ 3.2.5},
and are converted to JSON according to \code{schema}.}

\item{loc}{Alternative to \code{txt}. Location where input data is uploaded
and converted to internal server format.}

\item{location}{Legacy for \code{loc}}

\item{version}{Integer. JSON schema version number. There are currently two
schemas supported, versions \code{1} and \code{2}.}

\item{legacy}{Logical indicating whether legacy should be done.}
}
\value{
A JSON string containing a table with screening results
}
\description{
Screen growth curves according to JGZ guidelines
}
\note{
Deprecated for consistency. Function returns JSON, whereas all other
functions return the R object. The alternative \code{\link[=screen_growth]{screen_growth()}}
requests only results from screening. The alternative \code{\link[=custom_list]{custom_list()}}
produces the same list as \code{screen_curves}, but does not convert the
result to JSON.
}
\examples{
# # example json
# fn <- system.file("testdata", "client3.json", package = "james")
# fn <- system.file("testdata", "Laura_S_dev.json", package = "james")
#
# # first upload, then screen
# r1 <- jamesclient::upload_txt(fn)
# location <- jamesclient::get_url(r1, "location")
# location
# screen_curves(loc = location)
#
# # upload & screen
# screen_curves(fn)
}
