% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{screen_curves-deprecated}
\alias{screen_curves-deprecated}
\alias{screen_curves}
\title{Screen growth curves according to JGZ guidelines}
\usage{
screen_curves(
  txt = "",
  loc = "",
  location = "",
  format = "1.0",
  legacy = TRUE,
  ...
)
}
\arguments{
\item{txt}{A JSON string, URL or file with the data in JSON
format. The input data adhere to specification
\href{https://www.ncj.nl/themadossiers/informatisering/basisdataset/documentatie/?cat=12}{BDS JGZ 3.2.5},
and are converted to JSON according to \code{schema}.}

\item{loc}{Alternative to \code{txt}. Location where input data is uploaded.
Argument \code{loc} is deprecated and will disappear in Nov 2022; please
use \code{session} instead.}

\item{location}{Legacy for \code{loc}}

\item{format}{String. JSON data schema version number. There are currently
three schemas supported: \code{"1.0"}, \code{"1.1"}, \code{"2.0"} and \code{"3.0"}. Formats \code{"1.0"} and
\code{"1.1"} are included for backward compatibility only. Use \code{format = "3.0"}
for new applications.}

\item{legacy}{Logical indicating whether legacy should be done.}

\item{...}{Ignored}
}
\value{
A JSON string containing a table with screening results
}
\description{
Screen growth curves according to JGZ guidelines
}
\note{
Deprecated for consistency. Function returns JSON, whereas all other
functions return the R object. The alternative \code{\link[=screen_growth]{screen_growth()}}
requests only results from screening. The alternative \code{\link[=custom_list]{custom_list()}}
produces the same list as \code{screen_curves}, but does not convert the
result to JSON.
}
\examples{
# # example json
# fn <- system.file("testdata", "client3.json", package = "james")
# fn <- system.file("testdata", "Laura_S_dev.json", package = "james")
#
# # first upload, then screen
# r1 <- jamesclient::james_post(path = "data/upload", txt = fn)
# location <- jamesclient::get_url(r1, "location")
# location
# screen_curves(loc = location)
#
# # upload & screen
# screen_curves(fn)
}
