% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_curves.R
\name{screen_curves}
\alias{screen_curves}
\title{Screen growth curves according to JGZ guidelines}
\usage{
screen_curves(bds_data = NULL, ind_loc = NULL)
}
\arguments{
\item{bds_data}{A JSON string, URL or file. Optional.}

\item{ind_loc}{A url that points to the server location where the
data from a previous request to \code{convert_bds_ind()} are
stored. Optional. \code{ind_loc} takes priority over
\code{bds_data}.}
}
\value{
A JSON string
}
\description{
Screen growth curves according to JGZ guidelines
}
\examples{
\dontrun{
# example json
fn <- file.path(path.package("james"), "testdata", "client3.json")
js <- readLines(fn, warn = FALSE)

# first upload, then screen
r1 <- jamesclient::upload_bds(js)
location <- jamesclient::get_url(r1, "location")
location
screen_curves(ind_loc = location)

# implicit upload (slower)
screen_curves(bds_data = js)
}
}
