% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_curves.R
\name{screen_curves}
\alias{screen_curves}
\title{Screen growth curves according to JGZ guidelines}
\usage{
screen_curves(txt = "", loc = "", location = "", legacy = TRUE)
}
\arguments{
\item{txt}{A JSON string, URL or file with the data in JSON
format. The input data adhere to specification
\href{https://www.ncj.nl/themadossiers/informatisering/basisdataset/documentatie/?cat=12}{BDS JGZ 3.2.5},
and are converted to JSON according to \code{schema}.}

\item{loc}{Alternative to \code{txt}. Location where input data is uploaded
and converted to internal server format.}

\item{location}{Legacy for \code{loc}}

\item{legacy}{Logical indicating whether legacy should be done.}
}
\value{
A table with screening results
}
\description{
Screen growth curves according to JGZ guidelines
}
\examples{
\dontrun{
# example json
fn <- system.file("testdata", "client3.json", package = "james")
fn <- system.file("extdata", "smocc", "Laura_S.json", package = "jamestest")

# first upload, then screen
r1 <- upload_txt(fn)
location <- jamesclient::get_url(r1, "location")
location
screen_curves(loc = location)

# upload & screen
screen_curves(fn)
}
}
