% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_growth.R
\name{screen_growth}
\alias{screen_growth}
\title{Screen growth curves according to JGZ guidelines}
\usage{
screen_growth(txt = "", loc = "", version = 2L)
}
\arguments{
\item{txt}{A JSON string, URL or file with the data in JSON
format. The input data adhere to specification
\href{https://www.ncj.nl/themadossiers/informatisering/basisdataset/documentatie/?cat=12}{BDS JGZ 3.2.5},
and are converted to JSON according to \code{schema}.}

\item{loc}{Alternative to \code{txt}. Location where input data is uploaded
and converted to internal server format.}

\item{version}{Integer. JSON schema version number. There are currently two
schemas supported, versions \code{1} and \code{2}.}
}
\description{
Screen growth curves according to JGZ guidelines
}
\note{
\code{screen_growth} superseeds \code{screen_curves} and will
only return results from growth screening.
}
\examples{
host <- "http://localhost"
fn <- system.file("testdata", "client3.json", package = "james")

\dontrun{
# first upload, then screen
r1 <- jamesclient::upload_txt(fn, host = host)
location <- jamesclient::get_url(r1, "location")
location
screen_growth(loc = location)

# upload & screen
screen_growth(fn)
}
}
