% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_chart.R
\name{select_chart}
\alias{select_chart}
\title{Selects the growth chart}
\usage{
select_chart(ind = NULL, chartgrp = NULL, agegrp = NULL,
  sex = NULL, etn = NULL, ga = NULL, side = "hgt",
  language = "dutch")
}
\arguments{
\item{ind}{An S4 object of class \code{individual} containing
data of the individual.}

\item{chartgrp}{The chart group: \code{'nl2010'}, \code{'preterm'}, \code{'who'}
or \code{character(0)}}

\item{agegrp}{Either \code{'0-15m'}, \code{'0-4y'}, \code{'1-21y'},
\code{'0-21y'} or \code{'0-4ya'}. Age group \code{'0-4ya'}
provides the 0-4 chart with weight for age (design E).}

\item{sex}{Either \code{'male'} or \code{'female'}}

\item{etn}{Either \code{'netherlands'}, \code{'turkish'},
\code{'moroccan'} or \code{'hindustani'}}

\item{ga}{Gestionational age (in completed weeks)}

\item{side}{Either \code{'front'}, \code{'back'}, \code{'-hdc'} or
\code{'both'}}

\item{language}{Language: \code{'dutch'} or \code{'english'} (not
used)}
}
\value{
A list with elements \code{chartgrp}, \code{chartcode}
and \code{ga}
}
\description{
This function controls the behavior for selecting a specific growth
chart based on a combination of individual data and user settings.
The default behavior select preterm chart if gestional age is lower
or equal to 36 weeks, and determines the age group by the
maximum age found in the data.
}
\examples{
\dontrun{
cab <- groeivoorspeller::installed.cabinets
p <- cab[[1]][[1]]
select_chart(p)
}
}
\seealso{
\code{\link[chartcatalog]{create_chartcode}},
  \code{\link[minihealth]{individualAN-class}}
}
