% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_vwc.R
\name{select_vwc}
\alias{select_vwc}
\title{Calculates the recommended van Wiechen milestones based on age and past
milestones}
\usage{
select_vwc(
  txt = "",
  p = 90,
  n = 6,
  session = "",
  format = "1.0",
  loc = "",
  ...
)
}
\arguments{
\item{txt}{A JSON string, URL or file}

\item{p}{Reference percentile indicating the expected probability of a
positive van Wiechen outcome given child age. Higher values correspond with
easier items.}

\item{n}{Number of van Wiechen items to suggest. By default returns all items
within set probability limits.}

\item{session}{Optional session key if data is already uploaded to \code{sitehost}.}

\item{format}{String. JSON data schema version number. There are currently
three schemas supported: \code{"1.0"}, \code{"1.1"}, \code{"2.0"}, \code{"3.0"} and \code{"3.1"}.
Formats \code{"1.0"} and \code{"1.1"} are included for backward compatibility only.
Use \code{format = "3.1"} for new applications.}

\item{loc}{Deprecated. Use \code{session} instead.}

\item{...}{Ignored}
}
\value{
A string of milestones
}
\description{
Calculates the recommended van Wiechen milestones based on age and past
milestones
}
\examples{
fn <- system.file("testdata", "Laura_S.json", package = "james")
select_vwc(txt = fn, p = 50, n = 10)
}
\author{
Arjan Huizing 2025
}
