% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_vwc.R
\name{select_vwc}
\alias{select_vwc}
\title{Calculates the recommended van Wiechen milestones based on age and past
milestones}
\usage{
select_vwc(
  txt = "",
  refperc = 90,
  nsuggest = 6,
  session = "",
  format = "1.0",
  loc = "",
  percentiles = FALSE,
  ...
)
}
\arguments{
\item{txt}{A JSON string, URL or file with the data in JSON
format. The input data adhere to specification
\href{https://www.ncj.nl/themadossiers/informatisering/basisdataset/documentatie/?cat=12}{BDS JGZ 3.2.5},
and are converted to JSON according to \code{schema}.}

\item{refperc}{Reference percentile indicating the expected probability of a
positive van Wiechen outcome given child age. Higher values correspond with
easier items.}

\item{nsuggest}{Number of van Wiechen items to suggest.}

\item{session}{Alternative to \code{txt}. Session key where input data is uploaded
on \code{sitehost}.}

\item{format}{String. JSON data schema version number. There are currently
three schemas supported: \code{"1.0"}, \code{"1.1"}, \code{"2.0"} and \code{"3.0"}.
Formats \code{"1.0"} and \code{"1.1"} are included for backward compatibility only.
Use \code{format = "3.0"} for new applications.}

\item{loc}{Alternative to \code{txt}. Location where input data is uploaded.
Argument \code{loc} is deprecated and will disappear in Nov 2022; please
use \code{session} instead.}

\item{percentiles}{When \code{FALSE} (default), return a string vector containing
recommended van Wiechen items. When \code{TRUE}, returns a table with columns
\code{"item"} and \code{"D"} (D-score) and \code{"A"} (Age in months) columns for the
percentiles 2, 10, 50, 90 and 98.}

\item{...}{Ignored}
}
\value{
A vector or table.
}
\description{
Calculates the recommended van Wiechen milestones based on age and past
milestones
}
\examples{
fn <- system.file("testdata", "Laura_S_dev.json", package = "james")
vwc <- select_vwc(txt = fn, refperc = 50)
}
\author{
Arjan Huizing 2025
}
