% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_vwc.R
\name{select_vwc}
\alias{select_vwc}
\title{Calculates the recommended van Wiechen milestones based on age and past
milestones}
\usage{
select_vwc(
  txt = "",
  p = 90,
  n = 6,
  session = "",
  format = "1.0",
  loc = "",
  percentiles = FALSE,
  ...
)
}
\arguments{
\item{txt}{A JSON string, URL, or file with BDS data in JSON format. Data
should conform to the BDS JGZ 3.2.5 specification.}

\item{p}{Reference percentile indicating the expected probability of a
positive van Wiechen outcome given child age. Higher values correspond with
easier items.}

\item{n}{Number of van Wiechen items to suggest. By default returns all items
within set probability limits.}

\item{session}{Optional session key if data is already uploaded to \code{sitehost}.}

\item{format}{JSON schema version, e.g., \code{"3.0"}. Used when uploading.}

\item{loc}{Deprecated. Use \code{session} instead.}

\item{percentiles}{When \code{FALSE} (default), return a string vector containing
recommended van Wiechen items. When \code{TRUE}, returns a table with columns
\code{"item"} and \code{"D"} (D-score) and \code{"A"} (Age in months) columns for the
percentiles 2, 10, 50, 90 and 98.}

\item{...}{Ignored}
}
\value{
A vector or table.
}
\description{
Calculates the recommended van Wiechen milestones based on age and past
milestones
}
\examples{
fn <- system.file("testdata", "Laura_S.json", package = "james")
select_vwc(txt = fn, p = 50, n = 10)
}
\author{
Arjan Huizing 2025
}
