% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_chartcode.R
\name{validate_chartcode}
\alias{validate_chartcode}
\title{Validates a growth chart code}
\usage{
validate_chartcode(chartcode = "", ...)
}
\arguments{
\item{chartcode}{Chart code, typically something like \code{"NMAB"}}

\item{\dots}{Used for authentication}
}
\value{
A logical vector of with \code{length(chartcode)} elements.
}
\description{
This function checks whether \code{chartcode} is a available
in the chart library.
}
\seealso{
\code{\link[=list_charts]{list_charts()}}
}
